Moved core server logic to a base class
Moved the core server logic about start up and shutdown into a base class.
The desktop version has been ported over.
