+++ src/Compilers/Server/ServerShared/ServerShared.projitems
@@ -17,6 +17,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)ICompilerServerHost.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MetadataCache.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Connection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ServerClient.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ServerDispatcher.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)VisualBasicCompilerServer.cs" />
   </ItemGroup>
